requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
selenium==4.21.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.32.3
python-dateutil==2.9.0.post0
lxml==4.9.4